Add art_eeprom_files_updater program to update disciplining_config and temperature_table file exposed by the driver
